requirements.txt
dash
dash-bootstrap-components
pandas
autogluon.tabular
plotly
openpyxl
